Add sleep timeout after call to rsync
The rate limit for NCBI is 3 per second, so a timeout of 0.34 seconds is added after each rsync call so that we don't get our IP banned. In the future, it will also be good to change this to use the entrez toolkit, so that we can also add our email to the request, and then NCBI will email us if they have a problem with our usage of the service.