Download assemblies, unzip them, and generate BLASTDB with Python.
db_generator.py does the job of the old DBGenerator.bash. It uses the
BioPython Entrez module to download summaries (indexes) and assembly
data from a search term.

It includes a <unzip_all> function which recursively searches a directory for
gzip files and unzips them.

It includes a <directory_to_database> function which recursively
searches a directory for fasta files, then builds a BLAST nucleotide database
from them all.

As an example you can make a <data> directory and run <python3
db_generator.py> to download all the bradyrhizobia assemblies to the
data directory and build a database.